simulation/vacuum-robot-simulator/requirements.txt
pygame==2.1.3
numpy==1.21.6
pyserial==3.5